Refactor and remove GetAddressBuilder with AddressBalance
