--- src/node_api/high_level/mod.rs
@@ -1,8 +1,0 @@
-// Copyright 2021 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-//! Iota high level node APIs
-
-mod address;
-
-pub use self::address::*;
